+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   [#1466](https://github.com/greenbone/gsa/pull/1466) [#1467](https://github.com/greenbone/gsa/pull/1467)
 
 ### Changed
+- Changed filterdialog for tickets page [#1489](https://github.com/greenbone/gsa/pull/1489)
 - Restructured menu categories [#1481](https://github.com/greenbone/gsa/pull/1481)
 - Logout and usersettings link got a menu [#1481](https://github.com/greenbone/gsa/pull/1481)
 - Modified the BarChart's y-domain to avoid range [0,0]. [#1447](https://github.com/greenbone/gsa/pull/1447)